refactor: rename publishing persistence module
